[App not running] fixes
* Renamed main.py to app.py
* Renamed template directory to templates
* Added code to print server debug lines
